pyproject.toml
[tool.poetry]
name = "conversation-summarizer"
version = "0.1.0"
description = ""
authors = ["sahil <sahilofficial17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
groq = "^0.11.0"
streamlit = "^1.38.0"
fastapi = {extras = ["standard"], version = "^0.113.0"}
streamlit-audiorec = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
